fix: address the PR review comments
- asyncmy connections close via ensure_closed (close is the sync half)
- the pydantic suites import pymysql.cursors explicitly
- exact count(*) asserts on the shared tables became lower bounds, and
  the attrs/dataclass suites clean their enum-override rows like the
  other suites already did
- the omit_tc classes cleanup carries the missing dependency marker
- writeSqliteCall/writeSliceExpansion move to common.go as
  writeCursorCall (shared by the sqlite, turso, and MySQL drivers)
- CLAUDE.md drops a duplicated sentence; CONTRIBUTING, the docs landing
  description, and the drivers page mention MySQL where they still said
  postgres or sqlite only
