+++ test/driver_pymysql/dataclass/test_pymysql_dataclass_functions.py
@@ -787,7 +787,9 @@ class TestPymysqlDataclassFunctions:
     def test_count_mysql_types(self, pymysql_conn: pymysql.Connection) -> None:
         result = queries.count_mysql_types(conn=pymysql_conn)
 
-        assert result == 1
+        # The shared table may carry other files' rows; only a lower bound is safe.
+        assert result is not None
+        assert result >= 1
 
     @pytest.mark.dependency(name="PymysqlTestDataclassFunctions::all_cursor", depends=["PymysqlTestDataclassFunctions::count"])
     def test_all_mysql_types_cursor(self, pymysql_conn: pymysql.Connection, model: models.TestMysqlType) -> None:
@@ -1291,3 +1293,9 @@ class TestPymysqlDataclassFunctions:
         assert queries.count_mysql_types(conn=stub) is None
         assert queries_case.count_case_rows(conn=stub, id_=0) is None
         assert queries_enum_override.count_enum_override_by_moods(conn=stub, moods=[]) is None
+
+    @pytest.mark.dependency(depends=["PymysqlTestDataclassFunctions::insert_enum_override"])
+    def test_enum_override_cleanup(self, pymysql_conn: pymysql.Connection) -> None:
+        with pymysql_conn.cursor() as cur:
+            cur.execute("DELETE FROM test_enum_override WHERE id IN (%s, %s)", (ENUM_OVERRIDE_ID, ENUM_OVERRIDE_ID_2))
+        assert queries_enum_override.get_enum_override_mood(conn=pymysql_conn, id_=ENUM_OVERRIDE_ID) is None
